Adds a few small additions
Adds 25 MPH street signs

Adds the already sprited 40 MPH street signs to the map, had to code them.

Adds red aviators, and changes the yellow glasses to be named aviators

Renames Chicken Nugget to Chicken Wing for the time being.

renames the UN-AMERICAN CRISPS TO CHIPS

adds square infront of pizza

Changes all the listed names for Equipment vendors for clothing.

Changes the sprite for the medical bag on floor to be the proper color.