Additions to the code/map
adjust: fixes the cleaner phone number and removes the space so the number works

add: two new phones to the map to PD and Clinic front desks for phonecalls

add: places a bunch of papers listing all three emergency phone numbers for Clinic/PD/Cleaners (There are papers in the PD, Clinic. Giovanni mansion, Tower, Triad base, and the Anarch bar for Info on the numbers for easy access)

adjust: rotates the surgery bed.. IT WAS RIGHT THE FIRST TIME

adjust: locks the Tzim dungeon in the sewer, dif 7 (To prevent easy access but you can try to get in)

add: adds area zoning for the second Z-level to assist in Aerial bloodhunts on low rp Tzim

adjust: moves some areas to be more correct to how they should be, instead of having a split between the road it'll be more accurate to the actual start of the area.

add: therapy room for clinic

fixes: adds another button to the Cleaner burner so it works

add: places a Hammer and Sickle graffiti in the southern part of China town

delete: removes the extra set of doors in the museum