Generation nerf
This is mostly for test purposes. In tabletop generations do not give out additional health, and any blood pool they increase is by 1. This is supposed to level between higher and lower generation somewhat, there is no point staying high generation for roleplaying when lowering your generation is such a boost.